requirements.txt
affine==2.3.0
albumentations==1.1.0
annotated-types==0.5.0 
asn1crypto==0.24.0
attrs==21.4.0
awscli==1.21.3
Babel==2.5.1
beautifulsoup4==4.6.3
blivet==3.2.2
botocore==1.22.3
Brlapi==0.6.7
certifi==2021.10.8
cffi==1.11.5
chardet==3.0.4
click==8.0.4
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.3
configobj==5.0.6
configshell-fb==1.1.28
coverage==4.5.1
cryptography==3.2.1
cssselect==0.9.2
cupshelpers==1.0
cycler==0.11.0
dataclasses==0.8
dbus-python==1.2.4
decorator==4.4.2
dnf-etckeeper==0.0.0
dnspython==1.15.0
docutils==0.15.2
efficientnet-pytorch==0.6.3
ethtool==0.14
fail2ban==0.11.2
file-magic==0.3.0
gpg==1.13.1
gssapi==1.5.1
html5lib==0.999999999
humanize==0.5.1
idna==2.5
imageio==2.15.0
imgaug==0.4.0
importlib-metadata==4.8.3
importlib-resources==5.4.0
iniparse==0.4
iotop==0.6
ipaclient==4.9.6
ipalib==4.9.6
ipaplatform==4.9.6
ipapython==4.9.6
isc==2.0
Jinja2==2.10.1
jmespath==0.10.0
joblib==1.1.0
jsonpointer==1.10
jsonschema==2.6.0
jwcrypto==0.5.0
kerberos==1.3.0
kitchen==1.2.6
kiwisolver==1.3.1
kmod==0.1
libcomps==0.1.16
lit==13.0.1
lxml==4.2.3
Mako==1.0.6.dev0
MarkupSafe==0.23
matplotlib==3.3.4
meson==0.55.3
mlnx-tools==5.2.0
munch==2.5.0
netaddr==0.7.19
netifaces==0.10.6
networkx==2.5.1
nftables==0.1
nose==1.3.7
ntplib==0.3.3
numpy==1.19.5
opencv-python==4.5.5.64
opencv-python-headless==4.5.5.64
pandas==1.1.5
Paste==3.0.8
pciutils==2.3.6
pcp==5.0
perf==0.1
pexpect==4.3.1
Pillow==8.4.0
ply==3.9
pretrainedmodels==0.7.4
psycopg2==2.7.5
ptyprocess==0.5.2
pwquality==1.4.4
pyasn1==0.3.7
pyasn1-modules==0.1.5
pycairo==1.16.3
pycparser==2.14
pycups==1.9.72
pycurl==7.43.0.2
pydantic==2.1.1
pydantic-core==2.4.0
PyGObject==3.28.3
pyinotify==0.9.6
pyOpenSSL==19.0.0
pyparsing==2.1.10
pyparted==3.11.0
PySocks==1.6.8
python-augeas==0.5.0
python-cpio==0.1
python-dateutil==2.8.2
python-dmidecode==3.12.2
python-ldap==3.3.1
python-linux-procfs==0.6.3
python-meh==0.47.2
python-memcached==1.58
python-yubico==1.3.2
pytz==2017.2
pyudev==0.21.0
pyusb==1.0.0
PyWavelets==1.1.1
pyxdg==0.25
PyYAML==3.12
qrcode==5.1
qudida==0.0.4
rasterio==1.2.10
reportlab==3.4.0
requests==2.20.0
rhnlib==2.8.6
rpm==4.14.3
rsa==4.7.2
rtslib-fb==2.1.74
s3cmd==2.2.0
s3transfer==0.5.0
schedutils==0.6
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
scons==3.0.1
seaborn==0.11.2
segmentation-models-pytorch==0.2.1
selinux==2.9
sepolicy==1.1
setools==4.3.0
setproctitle==1.1.10
Shapely==1.8.1.post1
six==1.11.0
slip==0.6.4
slip.dbus==0.6.4
snuggs==1.4.7
sos==4.0
SSSDConfig==2.5.2
subscription-manager==1.28.13
suds-jurko==0.7.dev0
syspurpose==1.28.13
systemd-python==234
Tempita==0.5.1
threadpoolctl==3.1.0
tifffile==2020.9.3
timm==0.4.12
torch==1.8.1+cu111
torchaudio==0.8.1
torchvision==0.9.1+cu111
tqdm==4.63.0
typing_extensions==4.1.1
urlgrabber==4.1.0
urllib3==1.26.7
urwid==1.3.1
virtualenv==15.1.0
webencodings==0.5.1
zipp==3.6.0
